requirements.txt
streamlit
python-dotenv
openai
bcrypt
